Louis | Extracts classes into their own files
