+++ classes/bikewheelfem.py
@@ -8,30 +8,16 @@ import matplotlib.pyplot as pp
 import copy
 import re
 
-# Element type definitions
+from femsolution import FEMSolution
+from rigidbody import RigidBody
+from spokesection import SpokeSection
+from rimsection import RimSection
+from wheelgeometry import WheelGeometry
+from helpers import *
+
 EL_RIM = 1
 EL_SPOKE = 2
 TWOPI = 2*np.pi
-
-
-def skew_symm(v):
-    'Create a skew-symmetric tensor V from vector v such that V*u = v cross u.'
-
-    return np.matrix([[0,     v[2], -v[1]],
-                      [-v[2], 0,     v[0]],
-                      [v[1], -v[0],  0]])
-
-
-def interp_periodic(x, y, xx, period=TWOPI):
-    'Interpolate a periodic function with cubic spline matching slope at ends.'
-
-    # Pad data by wrapping beginning and end to match derivatives
-    x_pad = np.concatenate(([x[-1] - period], x, [x[0] + period, x[1] + period]))
-    y_pad = np.concatenate(([y[-1]], y, [y[0], y[1]]))
-
-    f_spline = interpolate.splrep(x_pad, y_pad)
-
-    return interpolate.splev(xx, f_spline)
 
 
 class BicycleWheelFEM:
@@ -501,379 +487,4 @@ class BicycleWheelFEM:
         self.soln_updated = False
 
 
-class BicycleWheelGeom:
-    'Geometric parameters including size and lacing pattern'
 
-    def parse_wheel_file(self, fname):
-
-        def keyword_rim(l, f):
-
-            # Default values
-            args = {'diam': 0.6, 'spacing': 'default', 'N': 36}
-
-            # Parse arguments
-            l_args = l.strip().split()
-            for c in l_args[1::]:
-                # Set option
-                key = c.split('=')[0]
-                arg = c.split('=')[1]
-                if key in args:
-                    args[key] = arg
-                else:
-                    raise ValueError('Invalid parameter: {:s}'.format(key))
-
-            self.d_rim = float(args['diam'])
-            N = int(args['N'])
-            self.n_rim_nodes = N
-
-            if args['spacing'] == 'default':
-
-                # Space rim nodes evenly
-                self.a_rim_nodes = np.linspace(0, 2*np.pi * (N-1)/N, N)
-                l = f.readline()
-
-            elif args['spacing'] == 'custom':
-
-                # Read rim node positions from input file
-                l = f.readline()
-
-                self.a_rim_nodes = np.zeros(N, dtype=np.float)
-                while len(l) > 0:
-
-                    l_rim_node = l.strip().split()
-
-                    if re.match('^\d+$', l_rim_node[0]):  # check if line is an integer
-                        self.a_rim_nodes[int(l_rim_node[0]) - 1] = np.radians(float(l_rim_node[1]))
-                        l = f.readline()
-                    elif l_rim_node[0][0] == '#':
-                        l = f.readline()
-                    else:
-                        break
-
-                # Make sure all angles are unique
-                if not len(self.a_rim_nodes) == len(set(self.a_rim_nodes)):
-                    raise ValueError('When using spacing=custom, all node angles must be defined')
-
-            else:
-                raise ValueError('Invalid value for parameter \'spacing\': ' + args['spacing'])
-
-            # Increment to the next line
-            return l
-
-        def keyword_hub(l, f):
-
-            # Default values
-            args = {'diam': 0.6, 'diam_drive': None, 'width': 0.035, 'width_drive': None, 
-                    'spacing': 'default', 'N': 36}
-
-            # Parse arguments
-            l_args = l.strip().split()
-            for c in l_args[1::]:
-                # Set option
-                key = c.split('=')[0]
-                arg = c.split('=')[1]
-                if key in args:
-                    args[key] = arg
-                else:
-                    raise ValueError('Invalid parameter: {:s}'.format(key))
-
-            self.d1_hub = float(args['diam'])
-            self.d2_hub = self.d1_hub
-            if args['diam_drive'] != None:
-                self.d2_hub = float(args['diam_drive'])
-
-            self.w1_hub = float(args['width'])
-            self.w2_hub = self.w1_hub
-            if args['width_drive'] != None:
-                self.w2_hub = float(args['width_drive'])
-
-            N = int(args['N'])
-            self.n_hub_nodes = N
-            self.s_hub_nodes = np.zeros(N, dtype=np.int8)
-
-            if args['spacing'] == 'default':
-
-                # Space rim nodes evenly
-                self.a_hub_nodes = np.linspace(0, 2*np.pi * (N-1)/N, N)
-                self.s_hub_nodes[::2] = 1    # Drive-side nodes
-                self.s_hub_nodes[1::2] = -1  # Left-side nodes
-                l = f.readline()
-
-            elif args['spacing'] == 'custom':
-
-                # Read hub node positions from input file
-                l = f.readline()
-
-                self.a_hub_nodes = np.zeros(N, dtype=np.float)
-
-                while len(l) > 0:
-
-                    l_hub_node = l.strip().split()
-
-                    if re.match('^\d+$', l_hub_node[0]):  # check if line is an integer
-                        i_node = int(l_hub_node[0]) - 1
-                        self.a_hub_nodes[i_node] = np.radians(float(l_hub_node[1]))
-
-                        if len(l_hub_node) > 2 and l_hub_node[2].upper() == 'D':
-                            self.s_hub_nodes[i_node] = 1  # Drive-side node
-                        else:
-                            self.s_hub_nodes[i_node] = -1  # Left-side node
-
-                        l = f.readline()
-                    elif l_hub_node[0][0] == '#':
-                        l = f.readline()
-                    else:
-                        break
-
-                # Make sure all angles are unique
-                if not len(self.a_hub_nodes) == len(set(self.a_hub_nodes)):
-                    raise ValueError('When using spacing=custom, all node angles must be defined')
-
-            else:
-                raise ValueError('Invalid value for parameter \'spacing\': ' + args['spacing'])
-
-            # Increment to the next line
-            return l
-
-        def keyword_lacing(l, f):
-            # Default values
-            args = {'pattern': 'default'}
-
-            # Parse arguments
-            l_args = l.strip().split()
-            for c in l_args[1::]:
-                # Set option
-                key = c.split('=')[0]
-                arg = c.split('=')[1]
-                if key in args:
-                    args[key] = arg
-                else:
-                    raise ValueError('Invalid parameter: {:s}'.format(key))
-
-            if args['pattern'] == 'custom':
-
-                # Read hub node positions from input file
-                l = f.readline()
-
-                self.lace_hub_n = np.array([], dtype=np.int32)
-                self.lace_rim_n = np.array([], dtype=np.int32)
-
-                while len(l) > 0:
-
-                    l_lace = l.strip().split()
-
-                    if re.match('^\d+$', l_lace[0]):  # check if line is an integer
-                        i_node_hub = int(l_lace[0])
-                        i_node_rim = int(l_lace[1])
-
-                        self.lace_hub_n = np.append(self.lace_hub_n, i_node_hub)
-                        self.lace_rim_n = np.append(self.lace_rim_n, i_node_rim)
-
-                        l = f.readline()
-                    elif l_lace[0][0] == '#':
-                        l = f.readline()
-                    else:
-                        break
-
-            self.n_spokes = len(self.lace_hub_n)
-
-            # Increment to the next line
-            return f.readline()
-
-        try:
-            with open(fname) as f:
-                l = f.readline()
-                while len(l) > 0:
-
-                    l_args = l.strip().split()
-
-                    if l_args[0].lower() == 'rim':
-                        l = keyword_rim(l, f)
-                    elif l_args[0].lower() == 'hub':
-                        l = keyword_hub(l, f)
-                    elif l_args[0].lower() == 'lacing':
-                        l = keyword_lacing(l, f)
-                    else:
-                        # Nothing interesting on this line. Skip it.
-                        l = f.readline()
-        except IOError as e:
-            print('I/O error({0}): {1}'.format(e.errno, e.strerror))
-            raise
-        except:
-            print('')
-            raise
-
-    def __init__(self, wheel_file=None, n_vec=np.array([0, 0, 1])):
-
-        print('# Initializing wheel geometry -----------')
-
-        # self.d_rim = d_rim   # rim diameter
-        # self.d_hub = d_hub   # hub diameter
-        # self.w1_hub = w1_hub # width from rim center to left flange
-        # self.w2_hub = w2_hub # width from rim center to right flange
-        self.n_vec = n_vec   # axial vector from hub center to drive side nut
-
-        if wheel_file is not None:
-            self.parse_wheel_file(wheel_file)
-
-
-class RimSection:
-    'Section definition for rim elements'
-
-    def __init__(self, area, I11, I22, I33, young_mod, shear_mod, K2=1.0, K3=1.0):
-        self.area = area
-        self.I11 = I11
-        self.I22 = I22
-        self.I33 = I33
-        self.young_mod = young_mod
-        self.shear_mod = shear_mod
-        self.K2 = K2
-        self.K3 = K3
-
-
-class SpokeSection:
-    'Section definition for spoke elements'
-
-    def __init__(self, d_spoke, young_mod):
-        self.young_mod = young_mod
-        self.area = np.pi / 4 * d_spoke**2
-
-
-class RigidBody:
-    'Set of nodes with constained DOFs'
-
-    def __init__(self, name, pos, nodes):
-
-        self.name = name
-
-        self.pos = np.array(pos)
-        self.nodes = np.array(nodes)
-        self.n_nodes = len(self.nodes)
-
-
-class FEMSolution:
-
-    def get_polar_displacements(self, node_id):
-        'Convert nodal displacements to polar form'
-
-        # Allow array input for node_id and/or dof
-        if not hasattr(node_id, '__iter__'):
-            node_id = [node_id]
-
-        u_rad = np.array([])
-        u_tan = np.array([])
-
-        for n in node_id:
-
-            x = np.array([self.x_nodes[n], self.y_nodes[n], self.z_nodes[n]])
-            n_tan = np.cross(np.array([0,0,1]),x)
-
-            u_rim = self.nodal_disp[6*n + np.arange(3)]
-
-            u_rad = np.append(u_rad, u_rim.dot(x) / np.sqrt(x.dot(x)))
-            u_tan = np.append(u_tan, u_rim.dot(n_tan) / np.sqrt(n_tan.dot(n_tan)))
-
-        u_rad = np.array(u_rad)
-        u_tan = np.array(u_tan)
-
-        return u_rad, u_tan
-
-    def get_deformed_coords(self, node_id, scale_rad, scale_tan):
-        'Get coordinates of nodes in scaled deformed configuration.'
-
-        # Allow array input for node_id and/or dof
-        if not hasattr(node_id, '__iter__'):
-            node_id = [node_id]
-
-        x_def = np.array([])
-        y_def = np.array([])
-        z_def = np.array([]) # TODO
-
-        u_rad, u_tan = self.get_polar_displacements(node_id)
-
-        for n in range(len(node_id)):
-            n_id = node_id[n]
-
-            x = np.array([self.x_nodes[n_id], self.y_nodes[n_id], self.z_nodes[n_id]])
-            n_tan = np.cross(np.array([0, 0, 1]), x)
-
-            u = scale_rad*u_rad * x/np.sqrt(x.dot(x)) + scale_tan*u_tan * n_tan
-
-            x_def = np.append(x_def, x[0] + u[0])
-            y_def = np.append(y_def, x[1] + u[1])
-
-        return x_def, y_def
-
-    def plot_deformed_wheel(self, scale_rad=0.1, scale_tan=0.0):
-
-        rim_nodes = [self.el_n1[e] for e in range(len(self.el_type)) if self.el_type[e] == EL_RIM]
-
-        u_rad, u_tan = self.get_polar_displacements(rim_nodes)
-
-        # Scale the largest displacement to a percentage of the rim radius
-        scale_rad = self.geom.d_rim/2 / max(np.abs(u_rad)) * scale_rad
-        scale_tan = self.geom.d_rim/2 / max(np.abs(u_rad)) * scale_tan
-
-        theta = self.geom.a_rim_nodes - np.pi/2
-
-        # Calculate coordinates in deformed configuration
-        theta_def = theta + scale_tan * u_tan / (self.geom.d_rim / 2)
-        r_def = (self.geom.d_rim / 2) + scale_rad * u_rad
-
-        theta_ii = np.linspace(-np.pi/2, 3*np.pi/2, 1000)
-
-        theta_interp = interpolate.interp1d(np.append(theta, theta[0] + 2*np.pi),
-                                            np.append(theta_def, theta_def[0] + 2*np.pi))
-        theta_def_ii = theta_interp(theta_ii)
-
-        r_def_ii = interp_periodic(theta, r_def, theta_ii)
-
-        # Plot undeformed rim
-        pp.plot(self.geom.d_rim/2 * np.cos(theta_ii),
-                self.geom.d_rim/2 * np.sin(theta_ii), 'k:')
-
-        # Plot deformed rim
-        pp.plot(r_def_ii * np.cos(theta_def_ii), r_def_ii * np.sin(theta_def_ii), 'k', linewidth=2.0)
-
-        # Plot spokes in deformed configuration
-        for e in [x for x in range(len(self.el_type)) if self.el_type[x] == EL_SPOKE]:
-            n_hub = self.el_n1[e]
-            n_rim = self.el_n2[e]
-
-            x_hub, y_hub = self.get_deformed_coords(n_hub, scale_rad, scale_tan)
-            x_rim, y_rim = self.get_deformed_coords(n_rim, scale_rad, scale_tan)
-
-            pp.plot([x_hub, x_rim], [y_hub, y_rim], 'k-')
-
-        # Axis properties
-        ax = pp.gca()
-        ax.set_ylim([-1.2*self.geom.d_rim/2, 1.2*self.geom.d_rim/2])
-        ax.set_xlim([-1.2*self.geom.d_rim/2, 1.2*self.geom.d_rim/2])
-
-        ax.set_aspect('equal', 'datalim')
-
-        ax.xaxis.set_major_locator(pp.NullLocator())
-        ax.yaxis.set_major_locator(pp.NullLocator())
-
-        return pp.gcf()
-
-    def __init__(self, fem):
-        self.updated = False
-
-        self.geom = copy.copy(fem.geom)
-
-        self.x_nodes = fem.x_nodes.copy()
-        self.y_nodes = fem.y_nodes.copy()
-        self.z_nodes = fem.z_nodes.copy()
-
-        self.el_type = fem.el_type.copy()
-        self.el_n1 = fem.el_n1.copy()
-        self.el_n2 = fem.el_n2.copy()
-
-        self.nodal_disp = []
-        self.nodal_rxn = []
-        self.dof_rxn = []
-        self.spoke_t = []
-        self.rim_n = []
-        self.rim_v = []
-        self.rim_m = []
