+++ example_stand_on_spokes.py
@@ -1,12 +1,13 @@
 #!/usr/bin/env python
 
-from bikewheelfem import *
+from classes.bikewheelfem import *
+from classes.wheelgeometry import WheelGeometry
 import matplotlib.pyplot as pp
 
 
 # Initialize wheel geometry from wheel files
-geom_36x3 = BicycleWheelGeom(wheel_file='wheel_36_x3.txt')
-geom_crow = BicycleWheelGeom(wheel_file='wheel_36_crowsfoot.txt')
+geom_36x3 = WheelGeometry(wheel_file='wheel_36_x3.txt')
+geom_crow = WheelGeometry(wheel_file='wheel_36_crowsfoot.txt')
 
 # Rim section and material properties
 r_sec = RimSection(area=82.0e-6,      # cross-sectional area
